Apply windowinsets
