+++ build.gradle
@@ -1,5 +1,6 @@
 buildscript {
     ext {
+        androidx_activity_version = "1.10.1"
         androidx_appcompat_version = "1.6.1"
         androidx_core_version = "1.9.0"
         androidx_exifinterface_version = "1.3.6"
