+++ sample/src/main/res/layout/activity_result.xml
@@ -2,6 +2,7 @@
 <LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/root_result"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical">
