+++ ucrop/build.gradle
@@ -38,6 +38,7 @@ android {
 }
 
 dependencies {
+    implementation "androidx.activity:activity:${androidx_activity_version}"
     implementation "androidx.appcompat:appcompat:${androidx_appcompat_version}"
     implementation "androidx.exifinterface:exifinterface:${androidx_exifinterface_version}"
     implementation "androidx.transition:transition:${androidx_transition_version}"
